+++ app/src/main/java/com/bunbeauty/fakelivestream/features/stream/view/StreamScreen.kt
@@ -22,9 +22,11 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -34,8 +36,14 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.style.TextOverflow.Companion.Ellipsis
 import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.compose.LifecycleEventEffect
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.navigation.NavHostController
 import com.bunbeauty.fakelivestream.R
 import com.bunbeauty.fakelivestream.features.stream.presentation.Stream
+import com.bunbeauty.fakelivestream.features.stream.presentation.StreamViewModel
 import com.bunbeauty.fakelivestream.features.stream.view.ui.AnimatedReaction
 import com.bunbeauty.fakelivestream.features.stream.view.ui.AvatarImage
 import com.bunbeauty.fakelivestream.features.stream.view.ui.CameraComponent
@@ -48,9 +56,44 @@ import com.bunbeauty.fakelivestream.ui.clickableWithoutIndication
 import com.bunbeauty.fakelivestream.ui.components.CachedImage
 import com.bunbeauty.fakelivestream.ui.components.ImageSource
 import com.bunbeauty.fakelivestream.ui.theme.FakeLiveStreamTheme
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
 
 @Composable
-fun StreamScreen(
+fun StreamScreen(navController: NavHostController) {
+    val viewModel: StreamViewModel = hiltViewModel()
+    val state by viewModel.state.collectAsStateWithLifecycle()
+    val onAction = remember {
+        { action: Stream.Action ->
+            viewModel.onAction(action)
+        }
+    }
+
+    val scope = rememberCoroutineScope()
+    LaunchedEffect(Unit) {
+        viewModel.event.onEach { event ->
+            when (event) {
+                Stream.Event.GoBack -> {
+                    navController.popBackStack()
+                }
+            }
+        }.launchIn(scope)
+    }
+    LifecycleEventEffect(Lifecycle.Event.ON_START) {
+        onAction(Stream.Action.Start)
+    }
+    LifecycleEventEffect(Lifecycle.Event.ON_STOP) {
+        onAction(Stream.Action.Stop)
+    }
+
+    StreamContent(
+        state = state.toViewState(),
+        onAction = onAction,
+    )
+}
+
+@Composable
+fun StreamContent(
     state: ViewState,
     onAction: (Stream.Action) -> Unit,
 ) {
@@ -563,7 +606,7 @@ private fun DirectBottomSheet(
 @Composable
 private fun StreamScreenPreview() {
     FakeLiveStreamTheme {
-        StreamScreen(
+        StreamContent(
             state = ViewState(
                 image = ImageSource.Res(R.drawable.img_default_avatar),
                 username = "long_user_name",
